requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# LangGraph + LangChain core
langgraph>=0.2.0
langchain>=0.2.0

# OpenRouter integration (this is what your code actually uses)
langchain-openrouter>=0.1.0  # instead of langchain-openai

pydantic>=2.7.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tenacity>=8.3.0
httpx>=0.27.0

pytest>=8.2.0
pytest-asyncio>=0.23.0